pyproject.toml
[tool.poetry]
name = "coursework_5_posgres_minin_es"
version = "0.1.0"
description = ""
authors = ["Евгений Минин <evgenminin1988@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
pytest = "^7.4.0"
coverage = "^7.2.7"
requests-mock = "^1.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
